Update: improve restore user files
- °customization.sh -> different actions for a CrossMix update compared to a first installation

- patch_ra_cfg.sh -> more flexible: can now manage spaces before and after "=" of each value and can manage quotes presence or not on the values. The target config file will keep its own format

- crossmix_update.sh -> more user datas restored (Retroarch cheats, database, RA filters, playlists, screenshots, thumbnails, shaders, RetroAchievements, PortMaster runtimes & themes, Tailscale config, syncthing config, ebooks, music player config, potential additional libs, CrossMix settings, Scraper Settings and themes, backgrounds and icons

+++ System/starts/°customization.sh
@@ -17,46 +17,50 @@ FW_patched_version=$(cat /usr/trimui/crossmix-version.txt)
 
 if [ "$version" != "$FW_patched_version" ]; then
 
-  # trimui_inputd update
+	if [ -f "/usr/trimui/crossmix-version.txt" ]; then
+		CrossMix_Update=1
+	else
+		CrossMix_Update=0
+	fi
 
-  Current_FW_Revision=$(grep 'DISTRIB_DESCRIPTION' /etc/openwrt_release | cut -d '.' -f 3)
-  Max_FW_Revision="20240503"
+	Current_FW_Revision=$(grep 'DISTRIB_DESCRIPTION' /etc/openwrt_release | cut -d '.' -f 3)
+	Max_FW_Revision="20240503"
 
-  if [ "$Current_FW_Revision" -le "$Max_FW_Revision" ]; then 
+	if [ "$Current_FW_Revision" -le "$Max_FW_Revision" ]; then
 
-      Sha_expected_1ms=c90b9fa722d745a7e827f38dbd409d3cd1ba56f5
-      Sha_expected_8ms=3f1b81d668c3f7de2cc0458502326a732d3cb0b2
-      Sha_expected_16ms=356b41b0be9d00f361e45303f41f5f1f337e6efc
-      Sha_1ms=$(sha1sum /mnt/SDCARD/System/resources/trimui_inputd/1ms | cut -d ' ' -f 1)
-      Sha_8ms=$(sha1sum /mnt/SDCARD/System/resources/trimui_inputd/8ms | cut -d ' ' -f 1)
-      Sha_16ms=$(sha1sum /mnt/SDCARD/System/resources/trimui_inputd/16s | cut -d ' ' -f 1)
-      if [ "$Sha_expected_1ms" = "$Sha_1ms" ] && [ "$Sha_expected_8ms" = "$Sha_8ms" ] &&
-        [ "$Sha_expected_16ms" = "$Sha_16ms" ]; then
+		Sha_expected_1ms=c90b9fa722d745a7e827f38dbd409d3cd1ba56f5
+		Sha_expected_8ms=3f1b81d668c3f7de2cc0458502326a732d3cb0b2
+		Sha_expected_16ms=356b41b0be9d00f361e45303f41f5f1f337e6efc
+		Sha_1ms=$(sha1sum /mnt/SDCARD/System/resources/trimui_inputd/1ms | cut -d ' ' -f 1)
+		Sha_8ms=$(sha1sum /mnt/SDCARD/System/resources/trimui_inputd/8ms | cut -d ' ' -f 1)
+		Sha_16ms=$(sha1sum /mnt/SDCARD/System/resources/trimui_inputd/16s | cut -d ' ' -f 1)
+		if [ "$Sha_expected_1ms" = "$Sha_1ms" ] && [ "$Sha_expected_8ms" = "$Sha_8ms" ] &&
+			[ "$Sha_expected_16ms" = "$Sha_16ms" ]; then
 
-          if [ ! -f /usr/trimui/bin/trimui_inputd.bak ]; then
-              mv /usr/trimui/bin/trimui_inputd /usr/trimui/bin/trimui_inputd.bak
-          fi
+			if [ ! -f /usr/trimui/bin/trimui_inputd.bak ]; then
+				mv /usr/trimui/bin/trimui_inputd /usr/trimui/bin/trimui_inputd.bak
+			fi
 
-          cp /mnt/SDCARD/System/resources/trimui_inputd/1ms /usr/trimui/bin/trimui_inputd.1ms
-          cp /mnt/SDCARD/System/resources/trimui_inputd/8ms /usr/trimui/bin/trimui_inputd.8ms
-          cp /mnt/SDCARD/System/resources/trimui_inputd/16ms /usr/trimui/bin/trimui_inputd.16ms
-          cp /mnt/SDCARD/System/resources/trimui_inputd/16ms /usr/trimui/bin/trimui_inputd
+			cp /mnt/SDCARD/System/resources/trimui_inputd/1ms /usr/trimui/bin/trimui_inputd.1ms
+			cp /mnt/SDCARD/System/resources/trimui_inputd/8ms /usr/trimui/bin/trimui_inputd.8ms
+			cp /mnt/SDCARD/System/resources/trimui_inputd/16ms /usr/trimui/bin/trimui_inputd.16ms
+			cp /mnt/SDCARD/System/resources/trimui_inputd/16ms /usr/trimui/bin/trimui_inputd
 
-          chmod +x /usr/trimui/bin/trimui_inputd.1ms
-          chmod +x /usr/trimui/bin/trimui_inputd.8ms
-          chmod +x /usr/trimui/bin/trimui_inputd.16ms
-          chmod +x /usr/trimui/bin/trimui_inputd
+			chmod +x /usr/trimui/bin/trimui_inputd.1ms
+			chmod +x /usr/trimui/bin/trimui_inputd.8ms
+			chmod +x /usr/trimui/bin/trimui_inputd.16ms
+			chmod +x /usr/trimui/bin/trimui_inputd
 
-          cp /mnt/SDCARD/System/resources/preload.sh /usr/trimui/bin/preload.sh
-      else
-          /mnt/SDCARD/System/usr/trimui/scripts/infoscreen.sh -m "Not able to update trimui_inputd, at least one new is missing or corrupted"
-      fi
+			cp /mnt/SDCARD/System/resources/preload.sh /usr/trimui/bin/preload.sh
+		else
+			/mnt/SDCARD/System/usr/trimui/scripts/infoscreen.sh -m "Not able to update trimui_inputd, at least one new is missing or corrupted"
+		fi
 
-  else
+	else
 
-      echo "recent firmware: trimui_inputd will not be updated"
+		echo "recent firmware: trimui_inputd will not be updated"
 
-  fi
+	fi
 
 	# Removing duplicated app
 	rm -rf /usr/trimui/apps/zformatter_fat32/
@@ -67,11 +71,11 @@ if [ "$version" != "$FW_patched_version" ]; then
 	rm -rf /swapfile
 	mv /bin/busybox.bak /mnt/SDCARD/System/bin 2>/dev/null
 	cp "/mnt/SDCARD/trimui/res/skin/bg.png" "/usr/trimui/res/skin/"
-	
+
 	# Increase alsa sound buffer
 	# cp "/mnt/SDCARD/System/usr/trimui/etc/asound.conf" "/etc/asound.conf"
 
-	# add Pl language + Fr and En mods
+	# add language files
 	if [ ! -e "/usr/trimui/res/skin/pl.lang" ]; then
 		cp "/mnt/SDCARD/trimui/res/lang/"*.lang "/usr/trimui/res/lang/"
 		cp "/mnt/SDCARD/trimui/res/lang/"*.short "/usr/trimui/res/lang/"
@@ -81,7 +85,7 @@ if [ "$version" != "$FW_patched_version" ]; then
 	# custom shutdown script for "Resume at Boot"
 	cp "/mnt/SDCARD/System/usr/trimui/bin/kill_apps.sh" "/usr/trimui/bin/kill_apps.sh"
 	chmod a+x "/usr/trimui/bin/kill_apps.sh"
-	
+
 	# fix retroarch path for PortMaster
 	cp "/mnt/SDCARD/System/usr/trimui/bin/retroarch" "/usr/bin/retroarch"
 	chmod a+x "/usr/bin/retroarch"
@@ -96,15 +100,27 @@ if [ "$version" != "$FW_patched_version" ]; then
 
 	# modifying FN cpu script (don't force slow cpu, only force high speed when FN is set to ON) (and we set it as default)
 	cp /mnt/SDCARD/System/usr/trimui/res/apps/fn_editor/com.trimui.cpuperformance.sh /usr/trimui/apps/fn_editor/com.trimui.cpuperformance.sh
-	cp /mnt/SDCARD/System/usr/trimui/res/apps/fn_editor/com.trimui.cpuperformance.sh /usr/trimui/scene/com.trimui.cpuperformance.sh
+
+	if [ "$CrossMix_Update" ]; then
+		if [ -f /usr/trimui/scene/com.trimui.cpuperformance.sh ]; then
+			cp /mnt/SDCARD/System/usr/trimui/res/apps/fn_editor/com.trimui.cpuperformance.sh /usr/trimui/scene/com.trimui.cpuperformance.sh
+		fi
+	else # in case of fresh install we set the default FN function on CPU Performance  Mode
+		cp /mnt/SDCARD/System/usr/trimui/res/apps/fn_editor/com.trimui.cpuperformance.sh /usr/trimui/scene/com.trimui.cpuperformance.sh
+	fi
+
+	# fix potential bad asound configuration
+		sed -i -e 's/period_size 2048/period_size 1024/' -e 's/period_size 4096/period_size 1024/' -e '/buffer_size 16384/d' "/etc/asound.conf"
 
 	# Apply default CrossMix theme, sound volume, and grid view
-	if [ ! -f /mnt/UDISK/system.json ]; then
-		cp /mnt/SDCARD/System/usr/trimui/scripts/MainUI_default_system.json /mnt/UDISK/system.json
-	else
-		/usr/trimui/bin/systemval theme "/mnt/SDCARD/Themes/CrossMix - OS/"
-		/usr/trimui/bin/systemval menustylel1 1
-		/usr/trimui/bin/systemval bgmvol 10
+	if [ ! "$CrossMix_Update" ]; then
+		if [ ! -f /mnt/UDISK/system.json ]; then
+			cp /mnt/SDCARD/System/usr/trimui/scripts/MainUI_default_system.json /mnt/UDISK/system.json
+		else
+			/usr/trimui/bin/systemval theme "/mnt/SDCARD/Themes/CrossMix - OS/"
+			/usr/trimui/bin/systemval menustylel1 1
+			/usr/trimui/bin/systemval bgmvol 10
+		fi
 	fi
 
 	# modifying performance mode for Moonlight
@@ -123,19 +139,23 @@ if [ "$version" != "$FW_patched_version" ]; then
 	"/mnt/SDCARD/Apps/SystemTools/Menu/THEME/Sort Themes Alphabetically.sh" -s
 
 	# Game tab by default
-	"/mnt/SDCARD/Apps/SystemTools/Menu/USER INTERFACE##START TAB (value)/Tab Game.sh" -s
+	if [ ! "$CrossMix_Update" ]; then
+		"/mnt/SDCARD/Apps/SystemTools/Menu/USER INTERFACE##START TAB (value)/Tab Game.sh" -s
+	fi
 
 	# Displaying only Emulators with roms
 	/mnt/SDCARD/Apps/EmuCleaner/launch.sh -s
 
 	################ Flash boot logo ################
-	CrossMixFWfile="/mnt/SDCARD/trimui/firmwares/MinFwVersion.txt"
-	Current_FW_Revision=$(grep 'DISTRIB_DESCRIPTION' /etc/openwrt_release | cut -d '.' -f 3)
-	Required_FW_Revision=$(sed -n '2p' "$CrossMixFWfile")
+	if [ ! "$CrossMix_Update" ]; then
+		CrossMixFWfile="/mnt/SDCARD/trimui/firmwares/MinFwVersion.txt"
+		Current_FW_Revision=$(grep 'DISTRIB_DESCRIPTION' /etc/openwrt_release | cut -d '.' -f 3)
+		Required_FW_Revision=$(sed -n '2p' "$CrossMixFWfile")
 
-	if ! [ "$Current_FW_Revision" -gt "$Required_FW_Revision" ]; then # on firmware hotfix 9 there is less space than before on /dev/mmcblk0p1 so we avoid to flash the logo
-		SOURCE_FILE="/mnt/SDCARD/Apps/BootLogo/Images/- CrossMix-OS.bmp"
-		"/mnt/SDCARD/Emus/_BootLogo/launch.sh" "$SOURCE_FILE"
+		if ! [ "$Current_FW_Revision" -gt "$Required_FW_Revision" ]; then # on firmware hotfix 9 there is less space than before on /dev/mmcblk0p1 so we avoid to flash the logo
+			SOURCE_FILE="/mnt/SDCARD/Apps/BootLogo/Images/- CrossMix-OS.bmp"
+			"/mnt/SDCARD/Emus/_BootLogo/launch.sh" "$SOURCE_FILE"
+		fi
 	fi
 fi
